+++ ios/Frameworks/KMPWorkManager.xcframework/ios-arm64_x86_64-simulator/KMPWorkManager.framework/Headers/KMPWorkManager.h
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a413e3750b6e1f96f600155705f9aadf7e84ccf652fc8cc25d0ed2de78b11e18
-size 297371
+oid sha256:1e04d553e826ff5db1d1bc663472ef544885b4b54f6f3d6900db80d23e1e965b
+size 297639